+++ dio/lib/src/dio/dio_for_native.dart
@@ -3,12 +3,12 @@ import 'dart:io';
 
 import '../adapter.dart';
 import '../cancel_token.dart';
+import '../dio_exception.dart';
 import '../dio_mixin.dart';
 import '../response.dart';
 import '../dio.dart';
 import '../headers.dart';
 import '../options.dart';
-import '../dio_error.dart';
 import '../adapters/io_adapter.dart';
 
 Dio createDio([BaseOptions? baseOptions]) => DioForNative(baseOptions);
@@ -49,8 +49,8 @@ class DioForNative with DioMixin implements Dio {
         queryParameters: queryParameters,
         cancelToken: cancelToken ?? CancelToken(),
       );
-    } on DioError catch (e) {
-      if (e.type == DioErrorType.badResponse) {
+    } on DioException catch (e) {
+      if (e.type == DioExceptionType.badResponse) {
         if (e.response!.requestOptions.receiveDataWhenStatusError == true) {
           final res = await transformer.transformResponse(
             e.response!.requestOptions..responseType = ResponseType.json,
@@ -140,7 +140,7 @@ class DioForNative with DioMixin implements Dio {
             await subscription.cancel();
           } finally {
             completer.completeError(
-              DioMixin.assureDioError(e, response.requestOptions),
+              DioMixin.assureDioException(e, response.requestOptions),
             );
           }
         });
@@ -153,7 +153,7 @@ class DioForNative with DioMixin implements Dio {
           completer.complete(response);
         } catch (e) {
           completer.completeError(
-            DioMixin.assureDioError(e, response.requestOptions),
+            DioMixin.assureDioException(e, response.requestOptions),
           );
         }
       },
@@ -162,7 +162,7 @@ class DioForNative with DioMixin implements Dio {
           await closeAndDelete();
         } finally {
           completer.completeError(
-            DioMixin.assureDioError(e, response.requestOptions),
+            DioMixin.assureDioException(e, response.requestOptions),
           );
         }
       },
@@ -180,7 +180,7 @@ class DioForNative with DioMixin implements Dio {
           await subscription.cancel();
           await closeAndDelete();
           if (e is TimeoutException) {
-            throw DioError.receiveTimeout(
+            throw DioException.receiveTimeout(
               timeout: timeout,
               requestOptions: response.requestOptions,
               error: e,