+++ dio/test/basic_test.dart
@@ -42,11 +42,13 @@ void main() {
     await expectLater(
       Dio().get('http://http.invalid'),
       throwsA(allOf([
-        isA<DioError>(),
-        (DioError e) =>
+        isA<DioException>(),
+        (DioException e) =>
             e.type ==
-            (isWeb ? DioErrorType.connectionError : DioErrorType.unknown),
-        if (!isWeb) (DioError e) => e.error is SocketException,
+            (isWeb
+                ? DioExceptionType.connectionError
+                : DioExceptionType.unknown),
+        if (!isWeb) (DioException e) => e.error is SocketException,
       ])),
     );
   });
@@ -63,7 +65,7 @@ void main() {
 
     await expectLater(
       dio.get('/test-timeout', cancelToken: token),
-      throwsA((e) => e is DioError && CancelToken.isCancel(e)),
+      throwsA((e) => e is DioException && CancelToken.isCancel(e)),
     );
   });
 
@@ -75,8 +77,8 @@ void main() {
       dio.get('/401'),
       throwsA(
         (e) =>
-            e is DioError &&
-            e.type == DioErrorType.badResponse &&
+            e is DioException &&
+            e.type == DioExceptionType.badResponse &&
             e.response!.statusCode == 401,
       ),
     );